norminette limpiada de redirecciones y syntax error. toca hacer norminette bien en ms_parser2.c y tambien modificar y arreglar el syntax error a ver si consigo que cuente varios pipes seguidos
