+++ libft/ft_arrsize.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/29 11:02:56 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/10 15:32:53 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/24 20:55:04 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
