+++ libft/ft_lstadd_back.c
@@ -6,14 +6,14 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/24 19:48:13 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/10 15:33:48 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/24 20:54:54 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 void	ft_lstadd_back(t_list **lst, t_list *new)
-{	
+{
 	t_list	*lstbck;
 
 	if (*lst)
