+++ libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/15 12:12:42 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/10 15:34:23 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/24 20:54:43 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 		while (l < n)
 		{
 			d[l] = s[l];
-				l++;
+			l++;
 		}
 	}
 	return (dst);
