+++ libft/ft_putchar_fd.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 09:55:01 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/10 15:34:31 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/24 20:54:30 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,4 +15,4 @@
 void	ft_putchar_fd(char c, int fd)
 {
 	write(fd, &c, 1);
-}	
+}
