+++ libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/13 12:05:22 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/10 15:35:00 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/24 20:54:05 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ size_t	ft_strlcpy(char *restrict dst, const char *restrict src, size_t dstsize)
 			dst[l] = src[l];
 			l++;
 		}
-	dst[l] = '\0';
+		dst[l] = '\0';
 	}
 	return (srclen);
 }
