+++ src/errors/ms_errors.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/05 20:41:03 by rabril-h          #+#    #+#             */
-/*   Updated: 2023/07/23 20:51:32 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/24 19:35:42 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,47 +37,4 @@ void	msh_errors_export(char *param, char *msg)
 	write(2, "minishell: export: `", 20);
 	write(2, param, ft_strlen(param));
 	write(2, msg, ft_strlen(msg));
-}
-
-int	msh_errors_syntax_extra(int flag, char *param)
-{
-	if (!param)
-		param = "";
-	if (flag == 0)
-		return (0);
-	ft_putstr_fd("minishell: syntax error near unexpected token `", 2);
-	ft_putstr_fd(param, 2);
-	ft_putstr_fd("'\n", 2);
-	g_return_status = 258;
-	return (1);
-}
-
-int	msh_errors_syntax(t_cmd *cmd, char *param)
-{
-	int	flag;
-
-	flag = 0;
-
-//dos pipes seguidos no funcionan como se deberia, deberia dar syntax error
-	while (cmd->next)
-	{
-		if (cmd->is_separator == 1 && cmd->next->is_separator == 1)
-		{
-			if (cmd->argv[0][0] != '|' && (cmd->next->argv[0][0] != '<'
-				|| cmd->next->argv[0][0] != '>' 
-				|| cmd->next->argv[0][0] != '|'))
-			{
-				flag = 1;
-				param = cmd->next->argv[0];
-				break ;
-			}
-		}
-		cmd = cmd->next;
-	}
-	if (cmd->is_separator == 1 && cmd->next == NULL)
-	{
-		flag = 1;
-		param = "newline";
-	}
-	return (msh_errors_syntax_extra(flag, param));
 }
