+++ src/parser/ms_parser.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ms_parser.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rabril-h <rabril-h@student.42barcelona.    +#+  +:+       +#+        */
+/*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/06 22:07:56 by rabril-h          #+#    #+#             */
-/*   Updated: 2023/07/23 19:28:32 by rabril-h         ###   ########.fr       */
+/*   Updated: 2023/07/24 20:58:41 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,87 +60,6 @@ int	msh_malformed_quotes(char *input)
 	return (0);
 }
 
-// put always qflag = 0;
-char	*msh_add_space_between_input(char *input, int qflag)
-{
-	int		i;
-	int		j;
-	int		count;
-	char	*result;
-
-	i = -1;
-	j = -1;
-	count = 0;
-	while (input[++i] != '\0')
-	{
-		qflag = msh_quote_flag(input[i], qflag);
-		if (qflag == 1 || qflag == 2)
-			continue ;
-		if (((input[i] == '>' && input[i + 1] != '>')
-				|| (input[i] == '<' && input[i + 1] != '<'))
-			&& input[i + 1] != ' ')
-			count ++;
-		else if (input[i] == '>' && input[i + 1] == '>' && input[i + 2] != ' ')
-		{
-			i ++;
-			count ++;
-		}
-		else if (input[i] == '<' && input[i + 1] == '<' && input[i + 2] != ' ')
-		{
-			i ++;
-			count ++;
-		}
-		else if ((input[i] != ' ' || input[i] != '>' || input[i] != '<')
-			&& (input[i + 1] == '>' || input[i + 1] == '<'))
-			count ++;
-	}
-	i = -1;
-	result = ft_calloc(sizeof(char), ft_strlen(input) + count + 1);
-	while (input[++i] != '\0')
-	{
-		qflag = msh_quote_flag(input[i], qflag);
-		if (qflag == 1 || qflag == 2)
-		{
-			result[++j] = input[i];
-			continue ;
-		}
-		if (((input[i] == '>' && input[i + 1] != '>')
-				|| (input[i] == '<' && input[i + 1] != '<'))
-			&& input[i + 1] != ' ')
-		{
-			result[++j] = input[i];
-			result[++j] = ' ';
-		}
-		else if (input[i] == '>' && input[i + 1] == '>' && input[i + 2] != ' ')
-		{
-			result[++j] = input[i];
-			result[++j] = input[++i];
-			result[++j] = ' ';
-		}
-		else if (input[i] == '<' && input[i + 1] == '<' && input[i + 2] != ' ')
-		{
-			result[++j] = input[i];
-			result[++j] = input[++i];
-			result[++j] = ' ';
-		}
-		else if ((input[i] != ' ' || input[i] != '>' || input[i] != '<')
-			&& (input[i + 1] == '>' || input[i + 1] == '<') 
-			&& input[i + 2] != ' ')
-		{
-			result[++j] = input[i];
-			result[++j] = ' ';
-		}
-		else
-		{
-			result[++j] = input[i];
-		}
-	}
-//	printf("result: %s\n", result);
-//	printf("count spaces = %d\n", count);
-	free(input);
-	return (result);
-}
-
 // ! Entry point
 char	*msh_sanitize_input(char *input)
 {
@@ -152,11 +71,7 @@ char	*msh_sanitize_input(char *input)
 		return (NULL);
 	}
 	trimmed_input = ft_strtrim(input, " ");
-	// ! comment for testing
-	// free(input);
+	free(input); // ! coment for testing
 	input = msh_clean_irrelveant_spaces_in_input(trimmed_input);
 	return (input);
 }
-
-
-
